remove app.js, move into home.js
